refactor: 잘못 나누어진 컴포넌트 제거
